교내공모전 병합4
교내공모전 병합4

+++ PageFinder/Assets/02.Scripts/Entity/Enemy/Data/EnemyData.cs
@@ -16,8 +16,8 @@ public class EnemyData : ScriptableObject
     public Enemy.AttackDistType attackDistType;
     public InkType inkType;
 
-    public int hp;
-    public int atk;
+    public float hp;
+    public float atk;
     public int cognitiveDist;
     public int inkTypeResistance;
     public int staggerResistance;
